rename errorText to placeholderText
thing is placeholder text can be derived from the errorCssClasses variable. So why introduce a new state.
Also, errorCssClasses can we derived from inputText somehow, if i only need error-border, but now since we want to re-render the component when the animation ends we need the state. Also there are many ways to do it. Like remove the class the in onAnimationEnd by using e.target.classList.remove and adding classes using document.querySelector in resetInputTextOnSave.
But, this is the most React way to doing I can think of.
Not the best, but "Good Enough (~James Shore)" for now.
It increases readability as well as understandability.